+++ common/platform.go
@@ -9,12 +9,3 @@ func IsAndoid() bool {
 func IsIOS() bool {
 	return runtime.GOOS == "ios"
 }
-func IsWindows() bool {
-	return runtime.GOOS == "windows"
-}
-func IsLinux() bool {
-	return runtime.GOOS == "linux"
-}
-func IsMac() bool {
-	return runtime.GOOS == "darwin"
-}
