+++ issue/issue_test.go
@@ -6,8 +6,9 @@ import (
 
 	"github.com/getlantern/fronted"
 	"github.com/getlantern/kindling"
+	"github.com/getlantern/radiance/api"
 	"github.com/getlantern/radiance/common"
-	"github.com/getlantern/radiance/user"
+
 	"github.com/stretchr/testify/require"
 )
 
@@ -19,8 +20,8 @@ func TestSendReport(t *testing.T) {
 		kindling.WithDomainFronting(f),
 		kindling.WithProxyless("api.iantem.io"),
 	)
-	userConfig := common.NewUserConfig("radiance-test", "","")
-	u := user.New(k.NewHTTPClient(), userConfig)
+	userConfig := common.NewUserConfig("radiance-test", "", "")
+	u := api.NewUser(k.NewHTTPClient(), userConfig)
 	reporter, err := NewIssueReporter(k.NewHTTPClient(), u, userConfig)
 	require.NoError(t, err)
 	// Grab a temporary directory
